dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev-requirements.in
#
attrs==21.2.0
    # via pytest
backports.entry-points-selectable==1.1.0
    # via virtualenv
certifi==2021.5.30
    # via
    #   -r requirements.txt
    #   syntropy-sdk
cfgv==3.3.1
    # via pre-commit
click==8.0.1
    # via -r requirements.txt
coverage==5.5
    # via -r dev-requirements.in
dataclasses==0.6
    # via -r requirements.txt
distlib==0.3.2
    # via virtualenv
filelock==3.0.12
    # via virtualenv
identify==2.2.14
    # via pre-commit
iniconfig==1.1.1
    # via pytest
nodeenv==1.6.0
    # via pre-commit
packaging==21.0
    # via pytest
platformdirs==2.3.0
    # via virtualenv
pluggy==1.0.0
    # via
    #   -r dev-requirements.in
    #   pytest
pre-commit==2.15.0
    # via -r dev-requirements.in
py==1.10.0
    # via
    #   -r dev-requirements.in
    #   pytest
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via -r dev-requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.txt
    #   syntropy-sdk
pyyaml==5.4.1
    # via
    #   -r requirements.txt
    #   pre-commit
randomize==0.14
    # via -r dev-requirements.in
six==1.16.0
    # via
    #   -r requirements.txt
    #   python-dateutil
    #   syntropy-sdk
    #   virtualenv
syntropy-sdk==0.3.0
    # via -r requirements.txt
toml==0.10.2
    # via
    #   pre-commit
    #   pytest
urllib3==1.26.6
    # via
    #   -r requirements.txt
    #   syntropy-sdk
virtualenv==20.7.2
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
